move restart to separate file
